refactor(server): /api/auth + 인증 미들웨어 분할 (Plan C #C9 12단계 — 마지막)
- src/server/routes/auth.ts 신규 (151줄): POST /api/auth + 4개 미들웨어 캡슐화
  (bearerAuth / sseTokenAuth / healSseKeyAuth / readOnlyGuard)
- setupAuth(api, ctx) 단일 진입점 — apiKey 유무에 따른 분기 일원화
- isValidSessionToken 헬퍼 제거 (sessionManager.validate 직접 사용)
- dashboard-api.ts에서 timingSafeEqual, Next 타입 import 제거
- dashboard-api.ts 402 → 290줄 (-112)

#C9 완료 — dashboard-api.ts 모든 라우트가 routes/* 모듈로 분할됨.
다음: #C12 createDashboardRoutes 시그니처 정리 (DashboardContext + authConfig 객체화)
